`date`

+++ recovery.sh
@@ -13,8 +13,9 @@ git show
 echo  "press enter key to continue"
 read
 
-
-cp ./init/.lldbinit ~/
+if [ `uname` == 'Darwin' ];then
+	cp ./init/.lldbinit ~/
+fi
 cp ./init/.vimrc ~/
 cp ./init/.vimrc.bundles ~/
 cp ./init/.gvimrc ~/
@@ -38,16 +39,22 @@ fi
 rm -rf ~/.vim/MySnippets
 cp -R ./init/vim/MySnippets ~/.vim/MySnippets
 
-defaults import com.googlecode.iterm2 ./init/com.googlecode.iterm2.plist
+if [ `uname` == 'Darwin' ];then
+	defaults import com.googlecode.iterm2 ./init/com.googlecode.iterm2.plist
+	mkdir -p ~/Library/Application\ Support/Sublime\ Text\ 3/Packages/User
+	cp -R ./init/sublime/User/* ~/Library/Application\ Support/Sublime\ Text\ 3/Packages/User/
+else
+	mkdir -p ~/.config/Sublime\ Text\ 3/Packages/User
+	cp -R ./init/sublime/User/* ~/.config/Sublime\ Text\ 3/Packages/User/
+fi
 
-mkdir -p ~/Library/Application\ Support/Sublime\ Text\ 3/Packages/User
-cp -R ./init/sublime/User/* ~/Library/Application\ Support/Sublime\ Text\ 3/Packages/User/
-
-mkdir -p  ~/Library/Developer/Xcode/UserData/
-rm -rf ~/Library/Developer/Xcode/UserData/CodeSnippets
-rm -rf ~/Library/Developer/Xcode/UserData/FontAndColorThemes
-rm -rf ~/Library/Developer/Xcode/UserData/KeyBindings
-cp -R ./init/Xcode/UserData/* ~/Library/Developer/Xcode/UserData/
+if [ `uname` == 'Darwin' ];then
+	mkdir -p  ~/Library/Developer/Xcode/UserData/
+	rm -rf ~/Library/Developer/Xcode/UserData/CodeSnippets
+	rm -rf ~/Library/Developer/Xcode/UserData/FontAndColorThemes
+	rm -rf ~/Library/Developer/Xcode/UserData/KeyBindings
+	cp -R ./init/Xcode/UserData/* ~/Library/Developer/Xcode/UserData/
+fi
 
 source ~/.zshrc
